shifted test to integration test
